Factored out trt_parameters
