+++ tensorflow/compiler/tf2tensorrt/convert/op_converter.h
@@ -21,6 +21,7 @@ limitations under the License.
 #include <vector>
 
 #include "absl/strings/str_format.h"
+#include "tensorflow/compiler/tf2tensorrt/convert/trt_parameters.h"
 #include "tensorflow/compiler/tf2tensorrt/convert/weights.h"
 #include "third_party/tensorrt/NvInfer.h"
